Refactored javascript controllers
